Bug isnt presennt on Mac? FInished health

+++ src/Units/Enemies/Cactus.cs
@@ -3,13 +3,13 @@ namespace MyGame
 {
 	public class Cactus : Unit
 	{
-		private int dmg;
+		private float dmg;
 		public Cactus (int dmg) : base ("cactusBmp", "walkingScrpt", -0.40f)
 		{
 			this.dmg = dmg;
 		}
 
-		public override int Dmg {
+		public override float Dmg {
 			get {
 				return dmg;
 			}
