+++ src/Units/Heros/Ranger.cs
@@ -7,7 +7,7 @@ namespace MyGame
 		{
 		}
 
-		public override int Dmg {
+		public override float Dmg {
 			get {
 				throw new NotImplementedException ();
 			}
